[Essential]Move DEBUG code paths from the main SceneDelegate to a new DebbugingSceneDelegate subclass to separate debig and test specific code from production
